valuation_api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
requests
python-dotenv
